Remove optional keys by default
The `bench_example` repo should be set so that a user can
run the default configuration with minimal changes.

Provide a link to the documentation in the config file
so that the user may find all available keys to be used
in the config file.

Addresses comments made by @ccarouge
https://github.com/CABLE-LSM/bench_example/pull/4#discussion_r1150065283